+++ wp-content/themes/intranet/library/Public/Site.php
@@ -30,3 +30,14 @@ if (!function_exists('municipio_table_of_contents_url')) {
         return $url;
     }
 }
+
+if (!function_exists('municipio_intranet_current_url')) {
+    /**
+     * Gets the current url
+     * @return string
+     */
+    function municipio_intranet_current_url()
+    {
+        return "//{$_SERVER['HTTP_HOST']}{$_SERVER['REQUEST_URI']}";
+    }
+}